+++ src/frontend/apps/desk/src/features/mail-domains/mailboxes/types.ts
@@ -1,5 +1,7 @@
 import { UUID } from 'crypto';
 
+import { ReactNode } from 'react';
+
 export interface MailDomainMailbox {
   id: UUID;
   local_part: string;
@@ -14,3 +16,19 @@ export type MailDomainMailboxStatus =
   | 'disabled'
   | 'pending'
   | 'failed';
+
+export interface ViewMailbox {
+  id: string;
+  first_name: string;
+  last_name: string;
+  local_part: string;
+  secondary_email: string;
+  status: MailDomainMailboxStatus;
+}
+
+export interface Step {
+  title: string;
+  content: ReactNode;
+  leftAction: ReactNode;
+  rightAction?: ReactNode;
+}